ellipsis-package {ellipsis}	R Documentation

ellipsis: Tools for Working with ...

Description

The ellipsis is a powerful tool for extending functions. Unfortunately this power comes at a cost: misspelled arguments will be silently ignored. The ellipsis package provides a collection of functions to catch problems and alert the user.

Author(s)

Maintainer: Hadley Wickham hadley@rstudio.com

Other contributors:

RStudio [copyright holder]

See Also

Useful links:

https://ellipsis.r-lib.org

https://github.com/r-lib/ellipsis

Report bugs at https://github.com/r-lib/ellipsis/issues

[Package ellipsis version 0.3.2 Index]