dots_empty {ellipsis}	R Documentation

Helper for consistent documentation

Description

Use @inheritParams ellipsis::dots_empty in your package to consistently document an unused ... argument.

Arguments

...
These dots are for future extensions and must be empty.

[Package ellipsis version 0.3.0 Index]